requirements.txt
aenum==2.2.1
asn1crypto==0.24.0
attrs==19.1.0
Automat==0.7.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
constantly==15.1.0
cryptography==2.7
cssselect==1.1.0
Faker==2.0.2
gremlinpython==3.4.3
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
isodate==0.6.0
lxml==4.4.1
parsel==1.5.2
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycparser==2.19
pycrypto==2.6.1
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pymongo==3.9.0
pyOpenSSL==19.0.0
python-dateutil==2.8.0
queuelib==1.5.0
requests==2.22.0
service-identity==18.1.0
six==1.12.0
text-unidecode==1.3
tornado==4.5.3
Twisted==19.7.0
urllib3==1.25.6
w3lib==1.21.0
zope.interface==4.6.0
